🌟 [Major]: Introducing the `Casing` module (#2)
## Description

This pull request introduces a new set of functions for handling various
string casing styles, removes an obsolete function, and adds
comprehensive tests for the new functionalities. The most important
changes include the addition of functions for converting, detecting, and
splitting casing styles, the removal of a deprecated test function, and
the inclusion of unit tests to ensure the correctness of the new
functions.

### New Functions for Casing Styles:

*
[`src/functions/public/ConvertTo-CasingStyle.ps1`](diffhunk://#diff-258b23af5ee150cb504dd0c5ae9ca077838800780474332a2bdca6831e5e7ba0R1-R71):
Added a function to convert strings to different casing styles,
supporting multiple styles such as `snake_case`, `UPPER_SNAKE_CASE`,
`kebab-case`, and more.
*
[`src/functions/public/Get-CasingStyle.ps1`](diffhunk://#diff-95b4ce37b8898c19567d7a2598dc12aa9e21c40f97459f2883b6dafc14c9b063R1-R99):
Added a function to detect the casing style of a given string.
*
[`src/functions/public/Split-CasingStyle.ps1`](diffhunk://#diff-fc6326239979c77a5e409ef744f5c3c80f517308d5fa4489f08c0a46d617bcd2R1-R101):
Added a function to split strings based on their casing style into an
array of words.

### Removal of Obsolete Function:

*
[`src/functions/public/Set-PSModuleTest.ps1`](diffhunk://#diff-488943d39ce767faa7f4dd40414635ec01da86c30b9ddd2074150da994093b9aL1-L22):
Removed the obsolete `Set-PSModuleTest` function, which was used for
testing purposes.

### Unit Tests for New Functions:

*
[`tests/Casing.Tests.ps1`](diffhunk://#diff-2f2303af7c9cde9c678233849880c4fbbd3a657ae65db7ee89ccdf0e659d373bR1-R124):
Added comprehensive unit tests for `Get-CasingStyle`,
`ConvertTo-CasingStyle`, and `Split-CasingStyle` functions to ensure
their correctness and reliability.
*
[`tests/PSModuleTest.Tests.ps1`](diffhunk://#diff-854f5db3c1afae9419d6a22487e65e9e83942e00d1f21a85d4c23424dd2f47cdL1-L5):
Removed the obsolete tests related to the `Set-PSModuleTest` function.

### Supporting Changes:

*
[`src/variables/private/Cases.ps1`](diffhunk://#diff-579288b54f74cabfb81b0f2728f31082cf93d67023b11cac51ff63bb7ad244e2R1-R10):
Added regex patterns for different casing styles to support the new
functions.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [x] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas